+++ cultura/views/espera.py
@@ -1,0 +1,7 @@
+from cultura.serializers.espera import CulturaEsperaAgrotoxicoSerializer
+
+from rest_framework.generics import CreateAPIView
+
+
+class CulturaEsperaAgrotoxicoAPIView(CreateAPIView):
+    serializer_class = CulturaEsperaAgrotoxicoSerializer
